backend/requirements.txt
# Flask and core
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
python-dotenv>=1.0.0

# HTTP client
requests>=2.31.0

# Ollama/LLM integration
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0
rank-bm25>=0.2.2

# Optional LLM providers (install as needed)
openai>=1.0.0  # For OpenAI GPT models
google-generativeai>=0.3.0  # For Google Gemini models
anthropic>=0.15.0  # For Anthropic Claude models

# Data processing
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0
